third_party/blink/web_tests/external/wpt/webrtc/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 17 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] RTCSessionDescription interface object length
  assert_equals: wrong value for RTCSessionDescription.length expected 1 but got 0
[FAIL] RTCSessionDescription interface: attribute type
  assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function set type() { [native code] }"
[FAIL] RTCSessionDescription interface: attribute sdp
  assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function set sdp() { [native code] }"
[FAIL] RTCIceCandidate interface: attribute relayProtocol
  assert_true: The prototype object must have a property "relayProtocol" expected true got false
[FAIL] RTCIceCandidate interface: attribute url
  assert_true: The prototype object must have a property "url" expected true got false
[FAIL] RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "relayProtocol" with the proper type
  assert_inherits: property "relayProtocol" not found in prototype chain
[FAIL] RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "url" with the proper type
  assert_inherits: property "url" not found in prototype chain
[FAIL] RTCPeerConnectionIceEvent interface: attribute url
  assert_true: The prototype object must have a property "url" expected true got false
[FAIL] RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent('ice') must inherit property "url" with the proper type
  assert_inherits: property "url" not found in prototype chain
[FAIL] RTCIceTransport interface: attribute component
  assert_true: The prototype object must have a property "component" expected true got false
[FAIL] RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "role" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "component" with the proper type
  assert_inherits: property "component" not found in prototype chain
[FAIL] RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "binaryType" with the proper type
  Right-hand side of 'instanceof' is not an object
[FAIL] RTCDTMFToneChangeEvent interface object length
  assert_equals: wrong value for RTCDTMFToneChangeEvent.length expected 1 but got 2
[FAIL] RTCErrorEvent must be primary interface of new RTCErrorEvent('error')
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Failed to construct 'RTCErrorEvent': 2 arguments required, but only 1 present."
[FAIL] Stringification of new RTCErrorEvent('error')
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Failed to construct 'RTCErrorEvent': 2 arguments required, but only 1 present."
[FAIL] RTCErrorEvent interface: new RTCErrorEvent('error') must inherit property "error" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Failed to construct 'RTCErrorEvent': 2 arguments required, but only 1 present."
Harness: the test ran to completion.
